Update User and ClashService classes, modify migration files
The User model now includes a method to validate users based on the balance. The ClashService is update to create a url and generate config files. The .gitignore file in the clash-config folder has been updated to exclude all but itself. Migration files for users_table and user_stats_table have been edited to reflect changes in traffic information and GitHub data structure. The console.php file has been cleaned and the web.php file updated to include a new route.

Signed-off-by: hldh214 <hldh214@gmail.com>