Add traffic usage to Dashboard and schedule stat update
The Dashboard component is updated to display the total amount of data traffic used by the authenticated user. For that, a 'formatBytes' function is added that formats the number of bytes into the suitable size. Moreover, the Laravel cron job scheduler is set to run 'app:update-stat-command' on an hourly basis which presumably updates the data usage statistics.

Signed-off-by: hldh214 <hldh214@gmail.com>