Update dashboard alert message
The wording of the alert message displayed on the Dashboard page has been adjusted. Instead of telling users their "account is not paid", the message now indicates that their "account is limited." This change aims to soften the tone of the message while still communicating the need for account funding.

Signed-off-by: hldh214 <hldh214@gmail.com>