Add user balance update command and traffic unpaid attribute
A new console command, UpdateBalanceCommand, is created to update user balance based on unpaid traffic data. The 'traffic_unpaid' attribute is introduced in the user model to track traffic that has not been charged to the user balance yet. The execution frequency of UpdateStatCommand is also adjusted to run every four hours for better performance.

Signed-off-by: hldh214 <hldh214@gmail.com>