Handle Github login exceptions and modify download filename
The GithubController now includes error handling for the Github login process. An exception is thrown if the Github account is already linked to another user, redirecting to the profile edit page. The ClashController, on the other hand, had a change in generating the filename for a downloaded file, now including the user's email.

Signed-off-by: hldh214 <hldh214@gmail.com>