Update stat calculation in UpdateStatCommand
The commit modifies the 'uplink' and 'downlink' stats calculation in the UpdateStatCommand.php file. It now multiplies the original values with the rate of each vmess server. This will result in more accurate stat updates reflecting the server's rate. A type hint comment for the VmessServer instance was also added.

Signed-off-by: hldh214 <hldh214@gmail.com>