Add payment processing and sponsorship webhook
Updated several models to include relationships for the implementation of payments, specifically the Payment, UserStat, VmessServer, and User models. Added a new endpoint in the GithubController to handle sponsor webhooks for processing payments. The application will respond to webhook events to create a payment, reflecting the user's Github sponsorship status. Modified the .env.example file to include the necessary configurations for webhook handling.

Signed-off-by: hldh214 <hldh214@gmail.com>