Add user balance update condition and user observer
An UserObserver has been created to update the 'last_settled_at' field whenever the user's balance is changed. Additionally, in the UpdateStatCommand, a check has been added to skip balance update if already updated for the day. Further changes include renaming `traffic_unpaid` comment to 'unsettled' in the users table migration and removing the 'uptime' field in the vmess_servers table migration.

Signed-off-by: hldh214 <hldh214@gmail.com>