Refactor ClashService and add low priority user handling
The ClashService has been refactored to move the vmess_servers parameter from the constructor to the genConf method. In addition, a property 'is_low_priority' was added to the User model, indicating users with negative balance. In the context of generating profile links, low priority users will only be assigned with one server.

Signed-off-by: hldh214 <hldh214@gmail.com>