Add trade status and redirect feature for Alipay payment
The code has been updated to include a periodic query to the Alipay gateway for the status of a transaction. Once the transaction shows a success status, the user is redirected to their profile edit page. Additionally, the payment ID is being passed in the query to enhance security and efficiency.

Signed-off-by: hldh214 <hldh214@gmail.com>