Add BalanceDetail model and update user stats handling
Introduced a new 'BalanceDetail' model and updated the database with a corresponding 'balance_details' migration table. The User stats handling code in the 'UpdateStatCommand' file was revised, improving the logic for calculating traffic deductions and incorporating the usage of the new BalanceDetail model. A new relationship method for balance details was added to the User model for easy retrieval of a user's balance details.

Signed-off-by: hldh214 <hldh214@gmail.com>