Add Alipay gateway and update payment process
This update introduces Alipay as a new payment gateway in the payment model. It also refactors the AlipayController for clearer separation of payment processes. Now, new order creation and scanning the QR code are split into two different routes and methods respectively. It also includes a check to prevent users from creating multiple unpaid payments. The frontend modules are adjusted accordingly to work with this new implementation.

Signed-off-by: hldh214 <hldh214@gmail.com>