Add server list display to Dashboard
Added a new function to render a list of servers in the Dashboard component. This was facilitated by fetching all VmessServer records in the DashboardController and passing them as a new prop to the Dashboard component. This feature will allow users to visualize available servers and their respective rates directly on the dashboard.

Signed-off-by: hldh214 <hldh214@gmail.com>