Refactor Alipay scan-to-pay UI and simplify trade status rendering
The commit simplifies the way trade statuses are handled and displayed in the Alipay Scan-to-Pay feature, moving away from function render calls to directly using object properties. Additionally, the layout of the Scan Component have been refactored to be more streamlined and efficient. This change should make the code more readable and maintainable.

Signed-off-by: hldh214 <hldh214@gmail.com>