Update clash-conf-template with detailed domain configurations
This update includes a significant expansion of the existing domain configurations. Numerous specific domain configurations for various services, included but not limited to, Apple, Microsoft, Alibaba, Tencent, and ByteDance have been added. This change is meant to improve the specificity and effectiveness of the network routing rules for these services.

Signed-off-by: hldh214 <hldh214@gmail.com>