Add Customer Service page and navigation
Created a new `CustomerServiceController` to handle the customer service page rendering and added the corresponding route. Also updated the application navigation to include a link to the new 'customer service' page both in the main and responsive navigations. Created the respective jsx view for the customer service page, providing contact information and messaging for customer inquiries.

Signed-off-by: hldh214 <hldh214@gmail.com>